Implement blueprint
