+++ app/templates/base.html
@@ -41,7 +41,7 @@
   <nav class="navbar navbar-expand-lg navbar-custom">
     <div class="container">
 
-      <a class="navbar-brand" href="{{ url_for('index') }}">
+      <a class="navbar-brand" href="{{ url_for('main.index') }}">
         HeartLink
       </a>
 
@@ -62,44 +62,44 @@
           <!-- Show different navbar links based on the user's login status --> 
           {% if is_logged_in %}
             <li class="nav-item">
-              <a class="nav-link {% if request.endpoint == 'home' %}active{% endif %}" href="{{ url_for('home') }}">
+              <a class="nav-link {% if request.endpoint == 'main.home' %}active{% endif %}" href="{{ url_for('main.home') }}">
                 Home
               </a>
             </li>
 
             <li class="nav-item">
-              <a class="nav-link {% if request.endpoint == 'matches' %}active{% endif %}" href="{{ url_for('matches') }}">
+              <a class="nav-link {% if request.endpoint == 'matches.matches' %}active{% endif %}" href="{{ url_for('matches.matches') }}">
                 Matches
               </a>
             </li>
 
             <li class="nav-item">
-              <a class="nav-link {% if request.endpoint == 'messages' %}active{% endif %}" href="{{ url_for('messages') }}">
+              <a class="nav-link {% if request.endpoint == 'messages.messages' %}active{% endif %}" href="{{ url_for('messages.messages') }}">
                 Messages
               </a>
             </li>
 
             <li class="nav-item">
-              <a class="nav-link {% if request.endpoint == 'profile' %}active{% endif %}" href="{{ url_for('profile') }}">
+              <a class="nav-link {% if request.endpoint == 'profiles.profile' %}active{% endif %}" href="{{ url_for('profiles.profile') }}">
                 Profile
               </a>
             </li>
 
             <li class="nav-item">
-              <a class="nav-link" href="{{ url_for('logout') }}">
+              <a class="nav-link" href="{{ url_for('auth.logout') }}">
                 Log out
               </a>
             </li>
 
           {% else %}
             <li class="nav-item">
-              <a class="nav-link {% if request.endpoint == 'index' %}active{% endif %}" href="{{ url_for('index') }}">
+              <a class="nav-link {% if request.endpoint == 'main.index' %}active{% endif %}" href="{{ url_for('main.index') }}">
                 Home
               </a>
             </li>
 
             <li class="nav-item d-flex align-items-center">
-              <a class="btn-login ms-3" href="{{ url_for('login') }}" id="openLogin">
+              <a class="btn-login ms-3" href="{{ url_for('auth.login') }}" id="openLogin">
                 Login
               </a>
             </li>
