+++ app/templates/myprofile.html
@@ -40,7 +40,7 @@
           >
           <div class="mt-3">
             <a 
-              href="{{ url_for('matches') }}" 
+              href="{{ url_for('matches.matches') }}" 
               class="btn-like"
             >
               Likes received❤️:
@@ -97,7 +97,7 @@
       </div>
 
       <div id="profileModal" class="profile-modal">
-        <form id="profileForm" class="profile-popup" action="{{ url_for('profile') }}" method="POST">
+        <form id="profileForm" class="profile-popup" action="{{ url_for('profiles.profile') }}" method="POST">
           <input type="hidden" name="csrf_token" value="{{ csrf_token() }}">
           <button id="closeProfile" class="close" type="button" aria-label="Close profile form">
             &times;
@@ -192,7 +192,7 @@
       </div>
 
       <div id="storyModal" class="profile-modal">
-        <form id="storyForm" class="profile-popup" method="POST" action="/update-story" enctype="multipart/form-data">
+        <form id="storyForm" class="profile-popup" method="POST" action="{{ url_for('profiles.update_story') }}" enctype="multipart/form-data">
           <input type="hidden" name="csrf_token" value="{{ csrf_token() }}">
           <button
             id="closeStory"
